--- config/passport.js
@@ -1,14 +1,0 @@
-// require('dotenv').config();
-// const passport = require('passport');
-// const Strategy = require('passport-github').Strategy;
-
-// passport.use(new Strategy({
-//     clientID: process.env[CLIENT_ID],
-//     clientSecret: process.env[CLIENT_SECRET],
-//     callbackURL: '/return'
-// },
-// function(accessToken, refreshToken, profile, cb) {
-
-// },
-// return cb(null, profile);
-// ));
