+++ views/index.handlebars
@@ -1,12 +1,10 @@
 <div class="container my-5">
     <div class="row">
         <div class="col">
-            {{> recentPodcasts-list }}
         </div>
     </div>
     <div class="row">
         <div class="col">
-            {{> updatedSubscriptions-list }}
         </div>
     </div>
     <div class="row mt-5">
