+++ views/layouts/main.handlebars
@@ -10,6 +10,7 @@
     <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
         <img src="/assets/img/..." height="80" class="d-inline-block align-top mx-auto" alt="">
         <a class="navbar-brand" href="#">Home</a>
+        <a class="btn gitbtn btn-success" href="#">Sign In with Github</a>
         <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
         <span class="navbar-toggler-icon"></span>
         </button>
