--- views/login.handlebars
@@ -1,29 +1,0 @@
-<div class="container-fluid">
-	<div class="row main-content bg-success text-center">
-		<div class="col-md-4 text-center company__info">
-			<span class="company__logo"><h2><span class="fa fa-android"></span></h2></span>
-			<h4 class="company_title">Your Company Logo</h4>
-		</div>
-		<div class="col-md-8 col-xs-12 col-sm-12 login_form ">
-			<div class="container-fluid">
-				<div class="row">
-					<h2>Sign In</h2>
-				</div>
-				<div class="row">
-					<form control="" class="form-group">	
-						<div class="row">
-							<button type="button" class="btn gitbtn btn-success">Sign In with Github</button>
-						</div>
-                        <div class="row">
-							<input type="checkbox" name="remember_me" id="remember_me" class="">
-							<label for="remember_me">Remember Me!</label>
-						</div>
-					</form>
-				</div>
-				<div class="row">
-					<p>Don't have an account? <a href="#">Register Here</a></p>
-				</div>
-			</div>
-		</div>
-	</div>
-</div>
