--- views/member.handlebars
@@ -1,1 +1,0 @@
-<h2>MEMBER PAGE</h2>
